+++ silo-core/test/foundry/Silo/borrow/Borrow.i.sol
@@ -547,19 +547,19 @@ contract BorrowIntegrationTest is SiloLittleHelper, Test {
 
         assertEq(
             SiloERC4626Lib._VIRTUAL_DEPOSIT_LIMIT - silo1TotalCollateral,
-            SiloERC4626Lib._VIRTUAL_DEPOSIT_LIMIT - silo1.total(ISilo.AssetType.Collateral),
+            SiloERC4626Lib._VIRTUAL_DEPOSIT_LIMIT - silo1.total(uint8(ISilo.AssetType.Collateral)),
             "limit for deposit"
         );
 
         assertEq(
             silo1.maxDeposit(borrower),
-            SiloERC4626Lib._VIRTUAL_DEPOSIT_LIMIT - silo1.total(ISilo.AssetType.Collateral),
+            SiloERC4626Lib._VIRTUAL_DEPOSIT_LIMIT - silo1.total(uint8(ISilo.AssetType.Collateral)),
             "can deposit when already borrowed"
         );
 
         assertEq(
             silo1.maxMint(borrower),
-            SiloERC4626Lib._VIRTUAL_DEPOSIT_LIMIT - silo1.total(ISilo.AssetType.Collateral),
+            SiloERC4626Lib._VIRTUAL_DEPOSIT_LIMIT - silo1.total(uint8(ISilo.AssetType.Collateral)),
             "can mint when already borrowed (maxMint)"
         );
     }
